Reworks CSP and CSS in demos; noop
