+++ demo/chess/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta http-equiv="content-security-policy" content="default-src 'self';">
     <link rel="stylesheet" href="index.css">
-    <script type="module" src="index.js"></script>
+    <script type="module" blocking="render" src="chess-piece.js"></script>
   </head>
   <body>
     <div id="container">
