+++ demo/hello-element.js
@@ -1,4 +1,5 @@
 import XElement from '../x-element.js';
+import styleSheet from './hello-element.css' with { type: 'css' };
 
 const logo = `\
   _________
@@ -10,38 +11,6 @@ const logo = `\
 
 class HelloElement extends XElement {
   static get styles() {
-    const styleSheet = new CSSStyleSheet();
-    styleSheet.replaceSync(`\
-      :host {
-        display: contents;
-      }
-
-      #container {
-        position: fixed;
-        --width: 150px;
-        --height: 150px;
-        --font-size: 13px;
-        font-weight: bold;
-        line-height: calc(var(--font-size) * 1.8);
-        font-size: var(--font-size);
-        top: calc(0px - var(--width) / 2);
-        left: calc(0px - var(--height) / 2);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: var(--width);
-        height: var(--height);
-        transform: translate(calc(0vw - var(--width)), 50vh) rotate(0deg);
-        opacity: 1;
-        transform-origin: center;
-        border-radius: 100vmax;
-        cursor: default;
-      }
-
-      #logo {
-        padding-bottom: var(--font-size);
-      }
-    `);
     return [styleSheet];
   }
 
@@ -73,5 +42,4 @@ class HelloElement extends XElement {
   }
 }
 
-
-customElements.define('hello-world', HelloElement);
+customElements.define('hello-element', HelloElement);
