+++ demo/uhtml/demo-uhtml.js
@@ -1,14 +1,8 @@
 import BaseElement from './base-element.js';
+import styleSheet from './demo-uhtml.css' with { type: 'css' };
 
 export default class DemoUhtml extends BaseElement {
   static get styles() {
-    const styleSheet = new CSSStyleSheet();
-    styleSheet.replaceSync(`
-      #container[emoji]::before {
-        content: " " attr(emoji);
-        font-size: 2rem;
-      }
-    `);
     return [styleSheet];
   }
