+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@netflix/x-element",
-  "description": "A dead simple starting point for custom elements.",
+  "description": "A dead simple basis for custom elements.",
   "version": "2.0.0-rc.11",
   "license": "Apache-2.0",
   "repository": "github:Netflix/x-element",
